Added nearest repeaters plugin
This patch adds the new POC nearest repeaters plugin.
It will fetch the closest N repeaters to a callsign's last
known GPS coordinates.  By default will return 2m repeaters,
but has the option to return other bands.

"n 3"  - return the closest 3 repeaters in 2m band
"n 1"  - return the 1 closest 2m band repeater
"n 70cm" - return the 1 closest 70cm band repeater
"n 3 70cm" - return the 3 closest 70cm band repeaters

This patch also updates the message thread to accept a list
of messages to send as a return from a plugin.   This results
in N aprs messages being sent as a reply to a command.

This was needed by the nearest plugin to allow it to return a repeater
per message as a response.