Migrate admin web out of aprsd.
This patch removes the admin command out of aprsd proper.
The admin interface will be an aprsd extension called
aprsd-admin-extension.

This is the start of the effort to trim down the base of APRSD
and the commands to server, healthcheck, list_plugins, list_extensions.

This will reduce the number of required packages that APRSD core needs.
This is useful for others using APRSD as their base to their
applications.